test(review-bundles): share pdf decode helpers
