+++ src/test/integration/review-bundles.test.mjs
@@ -3,8 +3,8 @@ import assert from "node:assert/strict";
 import fs from "node:fs";
 import path from "node:path";
 import os from "node:os";
-import zlib from "node:zlib";
 import { createTestContext } from "../helpers/server.js";
+import { decodePdfHexText, extractPdfFlateText } from "../helpers/pdf.js";
 
 const ctx = createTestContext(3071, 3070);
 let writeSyncDir, readSyncDir;
@@ -22,40 +22,6 @@ after(async () => {
 const callTool = (n, a) => ctx.callTool(n, a);
 const callWriteTool = (n, a) => ctx.callWriteTool(n, a);
 const waitForAsyncJob = (id, t) => ctx.waitForAsyncJob(id, t);
-
-function extractPdfFlateText(pdfBytes) {
-  const markerStart = Buffer.from("stream\n", "latin1");
-  const markerEnd = Buffer.from("\nendstream", "latin1");
-  const chunks = [];
-  let offset = 0;
-
-  while (offset < pdfBytes.length) {
-    const start = pdfBytes.indexOf(markerStart, offset);
-    if (start === -1) break;
-    const dataStart = start + markerStart.length;
-    const end = pdfBytes.indexOf(markerEnd, dataStart);
-    if (end === -1) break;
-    const compressed = pdfBytes.subarray(dataStart, end);
-    try {
-      chunks.push(zlib.inflateSync(compressed).toString("latin1"));
-    } catch {
-      // Non-flate or non-text stream; ignore.
-    }
-    offset = end + markerEnd.length;
-  }
-  return chunks.join("\n");
-}
-
-function decodePdfHexText(inflatedPdfText) {
-  const parts = [];
-  const re = /<([0-9A-Fa-f]+)>/g;
-  let match;
-  while ((match = re.exec(inflatedPdfText)) !== null) {
-    const hex = match[1].length % 2 === 0 ? match[1] : `0${match[1]}`;
-    parts.push(Buffer.from(hex, "hex").toString("latin1"));
-  }
-  return parts.join("");
-}
 describe("preview_review_bundle tool", () => {
   test("returns dry-run plan for outline profile", async () => {
     const text = await callTool("preview_review_bundle", {
